OneSystemToAnyOther task - TODO find bud

+++ Homeworks/03. CSharp-Advance/04. Numeral Systems/01.DecimalToBinary/DecimalToBinary.cs
@@ -6,7 +6,7 @@
     {
         public static void Main()
         {
-            long n = int.Parse(Console.ReadLine());
+            long n = long.Parse(Console.ReadLine());
 
             string result = ConvertDecimalToBinary(n);
